chore(server): move hash and jwt services
