+++ server/src/Infrastructure/Services/Jwt/JwtPayload.ts
@@ -1,0 +1,4 @@
+import type { JWTPayload } from "jose";
+export interface JwtPayload extends JWTPayload {
+  id: number;
+};
