+++ server/src/Tests/Unit/UseCases/Auth/ResetPasswordUseCase.spec.ts
@@ -3,7 +3,7 @@ import { describe, expect, it, vi, beforeEach } from "vitest";
 import type { User, Token } from "@prisma/client";
 
 import { Context } from "@Database/Core/Context";
-import { HashService } from "@Services/HashService";
+import { HashService } from "@Services/Hash/HashService";
 import { UnitOfWOrk } from "@Database/Core/UnitOfWork";
 import { UserFactory } from "@Database/Factories/UserFactory";
 import { TokenFactory } from "@Database/Factories/TokenFactory";
